SocketTest improvements
Create separate socket managers for client/server sides. Change
WinSocketManager to only call WSAStartup() and WSACleanup() once in the
processes's lifetime (in case multiple socket manager's are created).

Fix bug which sets the socket handle to -1 when it is closed. The handle
is size_t, which is unsigned.

Convey that the socket manager takes ownership of async sockets by
returning a weak reference to the socket in CreateAsync*Socket()
methods.

Have logging print to standard output if a LOG*() macro was called but
no logger has been created.